Extract `roxygen::to_code_action()`
